fix(options): split yaml row model shape
