+++ scripts/preview-freeze-check.mjs
@@ -52,9 +52,9 @@ const ALLOWED_PREVIEW_DRIFT = new Map([
     'original reference index.js vs current generated preview index.js',
     {
       reason:
-        'Generated Stitch Secondary preview JS intentionally carries the previously approved SVG-aware Stitch DOM helper, production parity/runtime harness wiring, schema-registered runtime surfaces, Reader/Video item-level session controls, the shared Reader/Video collapse action, Clipper removal of preview-incompatible visible cancel/comment/selection labels and the empty comment preview block, the inline video capture action plus readonly add-note trigger, persistent video control-bar auto-pause/screenshot preferences, the System interface theme preference, shared Clipper/Reader/Video export destination preview/dropdown content without visible destination carets, shared session width/height resize handle wiring, video fragment captures rendered as reader-style numbered text items with text-note placeholders, All in Ob production branding, QR-free task-success feedback toasts, the product-approved removal of the visible Settings sidebar title, production onboarding page routing from the Options resource link, release-hidden Experimental panel, Deep Research, AI timestamp, advanced video schema, prompt label/shortcut controls, and classifier placeholders, the performance branch classification timeout fallback enum, one-month Usage Dashboard history padding, Fragment modifier in-place chip sync, YAML button event containment, mouse-scroll-stable Stitch action buttons, render scroll restoration for Options actions, the video note-button switch promoted into the Video Prompt & Entry header with gray/green screenshot-dot guidance, default Domain Mappings restoration with an editable blank row only after user deletion plus inline-scrolling table chrome, Routing Rules canonical de-duplication, video timestamp screenshot toggle state rendered before the timestamp label, and the product-approved Reader/Video/task-success bottom-right floating runtime contract with no static task-success toast previews, compact feedback row, and Stitch-tokenized runtime toast root. Batch 2.5 additionally adds REST and vault local-folder File System Access controls, local-folder picker/remove/reauthorization actions, default-vault local-folder metadata syncing, storage schema wiring for local-folder rows, Options runtime routing/bootstrap preservation, button-action scroll preservation, and a task-success progress schema/content field absent from the frozen original reference. Milestone 7 relocates the preview-backed verification runtime from the retired source tree to tests/fixtures/options-preview while preserving the same frozen reference comparison contract. Milestone 10 splits the YAML Config Widget into adapter, controller, model conversion, validation presentation, field table rendering, and domain override rendering modules while preserving the production Stitch runtime contract. Milestone 13 splits the Stitch renderer into action, form, content, and node renderer modules while preserving the production Stitch runtime contract. Final verification stabilizes the preview bundle hash after the legacy fragment collectChanges return simplification without changing the frozen reference, freeze-check logic, or production behavior. The completion-audit format pass refreshes the generated JS hash after repo-wide Prettier normalization without changing the frozen reference or preview behavior. M3.1 splits the schema runtime renderer into context, widget lifecycle, hero, control, table, structural, element, node dispatch, and view modules while preserving the frozen reference and production Stitch runtime behavior. M5.4 refreshes the generated preview JS hash after the direct esbuild upgrade from ^0.23.0 to ^0.28.0. M3.3 refreshes the generated preview JS hash after production YAML sanitize guard narrowing changed bundled helper output while preserving the preview source, frozen reference, and Stitch runtime contract. M4 extracts native YAML widget DOM helpers into a thin typed module while preserving the preview source, frozen reference, and Stitch runtime contract. C1/C2 adds the config transfer sanitizer and StoredOptions root strip policy to the preview-backed Options runtime bundle while preserving the frozen reference, production copy/import behavior, and Stitch runtime contract. C3/C4 hardens classification and clip boundary schemas with strip semantics and bounded production meta fields, which changes only the schema runtime emitted into the preview bundle while preserving the frozen reference and visual/runtime preview contract. Plan 02 switches the preview YAML adapter to YamlConfigEditorWidgetAdapter while preserving widgetType: yaml-config, the Stitch runtime contract, and the original reference. Plan 09 strict typing repair refreshes the generated JS hash after exactOptionalPropertyTypes-safe YAML editor object construction without changing the frozen reference, freeze-check logic, or Stitch runtime behavior. YAML/Options retirement repair refreshes the generated JS hash after wiring the current YAML editor labels to injected i18n messages and default runtime message fallback while preserving the same Stitch widget contract. YAML runtime gap repair refreshes the generated JS hash after preserving disabled overrides for default custom fields and moving import-time YAML legacy-shape normalization ahead of StoredOptions schema validation while preserving the frozen reference and Stitch widget contract. YAML default custom field identity repair refreshes the generated JS hash after locking default custom field delete/rename controls and diffing default custom field overrides by stable baseline identity while preserving the frozen reference and Stitch widget contract. YAML default custom field cell-scope repair refreshes the generated JS hash after making non-owner content-type cells non-operable for default custom field rows while preserving explicit Add field behavior, baseline diff semantics, the frozen reference, and Stitch widget contract.',
+        'Generated Stitch Secondary preview JS intentionally carries the previously approved SVG-aware Stitch DOM helper, production parity/runtime harness wiring, schema-registered runtime surfaces, Reader/Video item-level session controls, the shared Reader/Video collapse action, Clipper removal of preview-incompatible visible cancel/comment/selection labels and the empty comment preview block, the inline video capture action plus readonly add-note trigger, persistent video control-bar auto-pause/screenshot preferences, the System interface theme preference, shared Clipper/Reader/Video export destination preview/dropdown content without visible destination carets, shared session width/height resize handle wiring, video fragment captures rendered as reader-style numbered text items with text-note placeholders, All in Ob production branding, QR-free task-success feedback toasts, the product-approved removal of the visible Settings sidebar title, production onboarding page routing from the Options resource link, release-hidden Experimental panel, Deep Research, AI timestamp, advanced video schema, prompt label/shortcut controls, and classifier placeholders, the performance branch classification timeout fallback enum, one-month Usage Dashboard history padding, Fragment modifier in-place chip sync, YAML button event containment, mouse-scroll-stable Stitch action buttons, render scroll restoration for Options actions, the video note-button switch promoted into the Video Prompt & Entry header with gray/green screenshot-dot guidance, default Domain Mappings restoration with an editable blank row only after user deletion plus inline-scrolling table chrome, Routing Rules canonical de-duplication, video timestamp screenshot toggle state rendered before the timestamp label, and the product-approved Reader/Video/task-success bottom-right floating runtime contract with no static task-success toast previews, compact feedback row, and Stitch-tokenized runtime toast root. Batch 2.5 additionally adds REST and vault local-folder File System Access controls, local-folder picker/remove/reauthorization actions, default-vault local-folder metadata syncing, storage schema wiring for local-folder rows, Options runtime routing/bootstrap preservation, button-action scroll preservation, and a task-success progress schema/content field absent from the frozen original reference. Milestone 7 relocates the preview-backed verification runtime from the retired source tree to tests/fixtures/options-preview while preserving the same frozen reference comparison contract. Milestone 10 splits the YAML Config Widget into adapter, controller, model conversion, validation presentation, field table rendering, and domain override rendering modules while preserving the production Stitch runtime contract. Milestone 13 splits the Stitch renderer into action, form, content, and node renderer modules while preserving the production Stitch runtime contract. Final verification stabilizes the preview bundle hash after the legacy fragment collectChanges return simplification without changing the frozen reference, freeze-check logic, or production behavior. The completion-audit format pass refreshes the generated JS hash after repo-wide Prettier normalization without changing the frozen reference or preview behavior. M3.1 splits the schema runtime renderer into context, widget lifecycle, hero, control, table, structural, element, node dispatch, and view modules while preserving the frozen reference and production Stitch runtime behavior. M5.4 refreshes the generated preview JS hash after the direct esbuild upgrade from ^0.23.0 to ^0.28.0. M3.3 refreshes the generated preview JS hash after production YAML sanitize guard narrowing changed bundled helper output while preserving the preview source, frozen reference, and Stitch runtime contract. M4 extracts native YAML widget DOM helpers into a thin typed module while preserving the preview source, frozen reference, and Stitch runtime contract. C1/C2 adds the config transfer sanitizer and StoredOptions root strip policy to the preview-backed Options runtime bundle while preserving the frozen reference, production copy/import behavior, and Stitch runtime contract. C3/C4 hardens classification and clip boundary schemas with strip semantics and bounded production meta fields, which changes only the schema runtime emitted into the preview bundle while preserving the frozen reference and visual/runtime preview contract. Plan 02 switches the preview YAML adapter to YamlConfigEditorWidgetAdapter while preserving widgetType: yaml-config, the Stitch runtime contract, and the original reference. Plan 09 strict typing repair refreshes the generated JS hash after exactOptionalPropertyTypes-safe YAML editor object construction without changing the frozen reference, freeze-check logic, or Stitch runtime behavior. YAML/Options retirement repair refreshes the generated JS hash after wiring the current YAML editor labels to injected i18n messages and default runtime message fallback while preserving the same Stitch widget contract. YAML runtime gap repair refreshes the generated JS hash after preserving disabled overrides for default custom fields and moving import-time YAML legacy-shape normalization ahead of StoredOptions schema validation while preserving the frozen reference and Stitch widget contract. YAML default custom field identity repair refreshes the generated JS hash after locking default custom field delete/rename controls and diffing default custom field overrides by stable baseline identity while preserving the frozen reference and Stitch widget contract. YAML default custom field cell-scope repair refreshes the generated JS hash after making non-owner content-type cells non-operable for default custom field rows while preserving explicit Add field behavior, baseline diff semantics, the frozen reference, and Stitch widget contract. YAML row-model shape repair refreshes the generated JS hash after extracting YAML field row aggregation/cell operability into a dedicated row model module while preserving runtime behavior and the frozen reference contract.',
       leftSha256: '9020ccbd91acd691eccd3fdf568b9a90efbddf0a35d79f36ef1caba702fa0c07',
-      rightSha256: '5b6f30c7aa9bec18d0d1d94e04b04680ec290714bc9a8c440dcec067aa36d6fb'
+      rightSha256: '5436a6896c81025f6de12dde4d207b7cfebdb647c86588ed73d1a72a48abc244'
     }
   ]
 ]);
